requirements.txt
contourpy==1.3.2
cupy-cuda12x==13.4.1
cycler==0.12.1
fastrlock==0.8.3
fonttools==4.58.1
kiwisolver==1.4.8
matplotlib==3.10.3
mpi4py==4.0.3
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.15.3
six==1.17.0
tzdata==2025.2
